+++ posthog/src/main/java/com/posthog/internal/PostHogRemoteConfig.kt
@@ -30,7 +30,7 @@ internal class PostHogRemoteConfig(
     private var featureFlags: Map<String, Any>? = null
     private var featureFlagPayloads: Map<String, Any?>? = null
 
-    // Decide v4 flags. These will later supersede featureFlags and featureFlagPayloads
+    // Flags v2 flags. These will later supersede featureFlags and featureFlagPayloads
     // But for now, we need to support both for back compatibility
     private var flags: Map<String, Any>? = null
     private var requestId: String? = null
@@ -122,7 +122,7 @@ internal class PostHogRemoteConfig(
                         if (hasFlags) {
                             if (config.preloadFeatureFlags) {
                                 if (distinctId.isNotBlank()) {
-                                    // do not process session replay from decide API
+                                    // do not process session replay from flags API
                                     // since its already cached via the remote config API
                                     executeFeatureFlags(distinctId, anonymousId, groups, onFeatureFlags, calledFromRemoteConfig = true)
                                 } else {
@@ -216,7 +216,7 @@ internal class PostHogRemoteConfig(
         }
 
         try {
-            val response = api.decide(distinctId, anonymousId = anonymousId, groups)
+            val response = api.flags(distinctId, anonymousId = anonymousId, groups)
 
             response?.let {
                 synchronized(featureFlagsLock) {
@@ -229,7 +229,7 @@ internal class PostHogRemoteConfig(
                         return@let
                     }
 
-                    val normalizedResponse = normalizeDecideResponse(it)
+                    val normalizedResponse = normalizeFlagsResponse(it)
 
                     if (normalizedResponse.errorsWhileComputingFlags) {
                         // if not all flags were computed, we upsert flags instead of replacing them
@@ -249,7 +249,7 @@ internal class PostHogRemoteConfig(
                         this.featureFlagPayloads = normalizedPayloads
                     }
 
-                    // only process and cache session replay config from decide API
+                    // only process and cache session replay config from flags API
                     // if not yet done by the remote config API
                     if (!calledFromRemoteConfig) {
                         processSessionRecordingConfig(it.sessionRecording)
@@ -369,13 +369,13 @@ internal class PostHogRemoteConfig(
         return parsedPayloads
     }
 
-    private fun normalizeDecideResponse(decideResponse: PostHogDecideResponse): PostHogDecideResponse {
-        val flags = decideResponse.flags
+    private fun normalizeFlagsResponse(flagsResponse: PostHogFlagsResponse): PostHogFlagsResponse {
+        val flags = flagsResponse.flags
         if (flags != null) {
             // This is a v4 response. This means that `featureFlags` and `featureFlagPayloads`
             // are not populated. We need to populate them with the values from the flags property.
             val newResponse =
-                decideResponse.copy(
+                flagsResponse.copy(
                     featureFlags = flags.mapValues { (_, value) -> value.variant ?: value.enabled },
                     featureFlagPayloads = flags.mapValues { (_, value) -> value.metadata.payload },
                 )
@@ -388,7 +388,7 @@ internal class PostHogRemoteConfig(
             }
             return newResponse
         }
-        return decideResponse
+        return flagsResponse
     }
 
     private fun loadFeatureFlagsFromCacheIfNeeded() {